+++ app/src/main/java/com/yeongjin/news/data/repository/NewsPagingRepository.kt
@@ -1,0 +1,14 @@
+package com.yeongjin.news.data.repository
+
+import com.yeongjin.news.data.remote.CategoryNewsPagingSource
+import com.yeongjin.news.data.remote.NewsPagingSource
+
+interface NewsPagingRepository {
+    fun getTopNewsList(country: String): NewsPagingSource
+
+    fun getCategoryNewsList(
+        country: String,
+        category: String
+    ): CategoryNewsPagingSource
+
+}
