+++ app/src/main/java/com/yeongjin/news/data/repositoryimpl/SavedNewsRepositoryImpl.kt
@@ -1,6 +1,7 @@
-package com.yeongjin.news.data.local.repository
+package com.yeongjin.news.data.repositoryimpl
 
 import com.yeongjin.news.data.local.datasource.SavedNewsLocalDataSource
+import com.yeongjin.news.data.repository.SavedNewsRepository
 import com.yeongjin.news.data.model.News
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
